Dataset architecture cleanup v3: surface attrs, decompose load_dataset, index-view subsets, DRY indices, extract halfsets
Phase 1: Surface image_stack attributes onto CryoEMDataset (process_images,
image_mask, data_multiplier, dataset_tilt_indices, tilt_particles). All 36
consumer references across 10 files now access these through CryoEMDataset
directly instead of reaching through cryo.image_stack.

Phase 2: Break load_dataset() into composable helpers (_create_image_stack,
_load_ctf_params, _load_poses, _apply_tilt_ctf_augmentation).

Phase 3: Replace _SubsampledImageStack with index-view pattern. Subset views
hold a reference to the original image_stack plus _subset_indices, with
_remap_generator translating backing indices to local 0..n-1 space.

Phase 4: Consolidate 6 near-identical _normalize_*_indices functions into
one canonical normalize_indices() in recovar/data_io/_index_utils.py.

Phase 5: Extract halfset-splitting logic (split_index_list, get_split_indices,
get_split_tilt_indices, figure_out_halfsets, etc.) to recovar/data_io/halfsets.py.
Re-exports in dataset.py preserve backward compatibility.

dataset.py: 1431 → 985 lines. All test-fast tests pass (2098 passed).

+++ recovar/reconstruction/regularization.py
@@ -28,7 +28,7 @@ def compute_batch_prior_quantities(rotation_matrices, translations, CTF_params, 
     return diag_mean
 
 def compute_prior_quantites(experiment_datasets, cov_noise, batch_size, for_whitening = False ):
-    bottom_of_fraction = jnp.zeros(experiment_datasets.volume_size, dtype = experiment_datasets[0].image_stack.dtype)
+    bottom_of_fraction = jnp.zeros(experiment_datasets.volume_size, dtype = experiment_datasets[0].dtype)
     for experiment_dataset in experiment_datasets:
         n_images = experiment_dataset.n_images
         # Compute the bottom of fraction.
